+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The rules for this file:
 - Added widget for Native Contacts (PR #59)
 - Added widget for Contacts within cutoff (PR #60)
 - Added widget for number of Hydrogen bonds (PR #61)
+- Added widget for Helix Analysis (PR #62)
 
 ### Fixed